Implementation of the matplotlib cpp library
Implemented a c++ library for matplotlib as alternative to FLTK. Can be enabled and disabled using -DUSING_FLTK=1 or DUSING_MATPLOTLIB =1  respectively. If both are unset no plotting system will be compiled

Massive changes to the j_input.cpp to improve reading speed and general logic of reading a file with regards to abstract positioning of controls, subcircuits and main circuit parts indroduced by WRSpice.